chore: export logout mutation from component
